requirements.txt
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.0
pyparsing==3.0.8
pytesseract==0.3.9